pyproject.toml
[tool.poetry]
name = "krxns"
version = "0.1.0"
description = ""
authors = ["stefanpate <stefanpate94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.0"
matplotlib = "^3.9.2"
networkx = "^3.3"
ipykernel = "^6.29.5"
rdkit = "^2024.3.5"
pandas = "^2.2.2"
pyyaml = "^6.0.2"
pathlib = "^1.0.1"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
